requirements.txt
langchain_community
langchain_openai
psycopg2-binary
python-dotenv
